+++ moira_client/models/__init__.py
@@ -1,0 +1,4 @@
+from .trigger import Trigger
+from .contact import Contact
+from .pattern import Pattern
+from .subscription import Subscription