+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.2
+- TriggerManager is_exist function added
+- TriggerManager get_non_existent function added 
+
 # 0.1.1
 - Trigger exist check added